+++ pkg/template/catalog/error.go
@@ -1,0 +1,9 @@
+package catalog
+
+import (
+	"github.com/giantswarm/microerror"
+)
+
+var executionFailedError = &microerror.Error{
+	Kind: "executionFailedError",
+}